day010_01_Win32API_Collider_2

+++ Client/CPathMgr.cpp
@@ -15,7 +15,7 @@ void CPathMgr::init()
 	GetCurrentDirectory(255, m_szContentPath); // exe 실행되면 실행된 자신의 위치를 저장하는데 그 정보를 Get하는 함수
 
 	// 문자열길이
-	int iLen = wcslen(m_szContentPath);
+	int iLen = (int)wcslen(m_szContentPath);
 
 	// 상위폴더로
 	for (int i = iLen; 0 <= i; --i)
